Feature/image input (#13)
* add basic image input

* increment version and add sample image input code

* format code

+++ aio_straico/async_client.py
@@ -122,6 +122,7 @@ class AsyncStraicoClient:
         message,
         *,
         files: [Path | str] = [],
+        images: [Path | str] = [],
         youtube_urls: [str] = [],
         temperature: float = None,
         max_tokens: float = None,
@@ -133,8 +134,8 @@ class AsyncStraicoClient:
             model = model["model"]
         elif type(model) == Model:
             model = model.model
-
-        if len(files) > 0 or len(youtube_urls) > 0:
+        v = None
+        if len(files) > 0 or len(youtube_urls) > 0 or len(images) > 0:
             if is_listable_not_string(files) and len(files) > 4:
                 raise Exception(
                     f"Too many attached files. API is limited to 4 File attachments"
@@ -143,20 +144,25 @@ class AsyncStraicoClient:
                 raise Exception(
                     f"Too many youtube_urls files. API is limited to 4 Youtube URL attachments"
                 )
+            if is_listable_not_string(images) and len(images) > 4:
+                raise Exception(
+                    f"Too many image files. API is limited to 4 Image URL attachments"
+                )
             v = 1
-        if isinstance(model, tuple) or isinstance(model, list):
-            v = 1
-            new_model = []
-            for m in model:
-                if isinstance(m, dict) and "model" in m:
-                    new_model.append(m["model"])
-                elif isinstance(m, Model):
-                    new_model.append(m.model)
-                else:
-                    new_model.append(m)
-            model = new_model
-        else:
-            v = 0
+        if v is None:
+            if isinstance(model, tuple) or isinstance(model, list):
+                v = 1
+                new_model = []
+                for m in model:
+                    if isinstance(m, dict) and "model" in m:
+                        new_model.append(m["model"])
+                    elif isinstance(m, Model):
+                        new_model.append(m.model)
+                    else:
+                        new_model.append(m)
+                model = new_model
+            else:
+                v = 0
 
         if v == 0:
             response = await aio_prompt_completion0(
@@ -176,22 +182,44 @@ class AsyncStraicoClient:
             if isinstance(files, str):
                 files = [files]
 
+            if isinstance(images, str):
+                images = [images]
+
             file_urls = []
             for file in files:
                 if isinstance(file, str):
                     if not file.strip().lower().startswith("http"):
                         file_path = Path(file)
-                        if file_path.exist():
+                        if file_path.exists() and file_path.is_file():
                             file_url = await self.upload_file(file)
                         else:
                             raise Exception(f"Unknown file {file}")
                     else:
                         file_url = file
-                elif isinstance(file, Path):
+                elif isinstance(file, Path) and file.exists() and file.is_file():
                     file_url = await self.upload_file(file)
                 else:
                     raise Exception(f"Unknown file type {type(file)} for file {file}")
                 file_urls.append(file_url)
+
+            images_urls = []
+            for image in images:
+                if isinstance(image, str):
+                    if not image.strip().lower().startswith("http"):
+                        file_path = Path(image)
+                        if file_path.exists() and file_path.is_file():
+                            image_url = await self.upload_file(image)
+                        else:
+                            raise Exception(f"Unknown file {image}")
+                    else:
+                        image_url = image
+                elif isinstance(image, Path) and image.exists() and image.is_file():
+                    image_url = await self.upload_file(image)
+                else:
+                    raise Exception(
+                        f"Unknown file type {type(image)} for image {image}"
+                    )
+                images_urls.append(image_url)
 
             response = await aio_prompt_completion1(
                 self._session,
@@ -200,6 +228,7 @@ class AsyncStraicoClient:
                 model,
                 message,
                 file_urls=file_urls,
+                images=images_urls,
                 youtube_urls=youtube_urls,
                 display_transcripts=display_transcripts,
                 temperature=temperature,
@@ -238,6 +267,11 @@ class AsyncStraicoClient:
             "kt": "text/x-kotlin",
             "xml": "application/xml",
             "ts": "application/typescript",
+            "png": "image/png",
+            "jpg": "image/jpeg",
+            "jpeg": "image/jpeg",
+            "webp": "image/webp",
+            "gif": "image/gif",
         }
         if not file_to_upload.exists():
             raise Exception(f"Cannot find file {file_to_upload}")
